Dependency Injection and Controller implementation
Configured an scoped service that maps the ICommandRepo interface to the MockRepo class.

Created the implementation for 2 HTTP GET requests:

on route api/commands all elements of the mock repository are returned with a 200 OK status

on route api/commands/{id} a harcoded single element on the mock repository is returned. In the future, this action will return an element that matches the id passed on the route.